check for unordered input data
LIBLINEAR (as well as LibSVM) requires input data (labels) to be ordered
by indices in ascending order.
However, it seems to be a common mistake to use LIBLINEAR without
ordered data. It leads to a weird behaviour in the C++ version and
usually ArrayIndexOutOfBounds exceptions in the Java version.

With this commit, the Java version of LIBLINEAR checks if the labels are
ordered correctly and throws an InvalidInputDataException in the other
case.