add an interface for feature nodes
This makes it possible to use a customized feature node implementation.
For example for features that always have a value of 1.0 and hence can
save the 8 bytes of memory.

Benchmarks show that there’s no performance penalty with modern JIT
compilers, when using the getIndex() and getValue() method calls instead
of directly accessing the properties.